+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- `useDeviceHook` for responsiveness.
+- `withDevice` HOC for responsiveness.
 - Mobile behavior for the `FilterBar` component.
 
 ## [9.105.6] - 2020-01-27